requirements.txt
# Core dependencies
pygame>=2.5.0
Pillow>=10.0.0
PyYAML>=6.0
python-vlc>=3.0.0

# GPIO support (Raspberry Pi only)
RPi.GPIO>=0.7.1

# Async support
asyncio>=3.4.3

# Development dependencies
pytest>=7.0.0
black>=23.0.0
pylint>=2.17.0
flake8>=6.0.0
mypy>=1.0.0
pytest-cov>=4.0.0